make sure to create the ~/.micas file if it doesn't already exist. Fixes #51
This issue was not fixed fully, there was instance of code in `routes.py` which ran before the `~/.micas` file could be created. Therefore, the code was decoupled a bit. Functions for validating locations and getting the UID were decoupled and the place were the `~/.micas` file is created was centralized to one location.